feat [ui]: Allow custom label breadcrumbs
