+++ ui/pages/actions/pkgs/deb-repo.mjs
@@ -27,11 +27,17 @@ const Docs = ({ field }) => (
 
 const fields = {
   basics: {
+    Package: 'The name to use for the package',
     Description: 'A one-liner short description',
     Version: 'The version number of the package',
     Revision: 'The revision number of the package',
   },
   advanced: {
+    Section: 'The Debian section the package belongs to',
+    Priority: 'The Debian priority code for this package',
+    Architecture: 'The Debian architecture code for the package',
+    Essential: 'Setting this to yes means the package cannot be removed',
+    'Installed-Size': 'Estimate of the disk space reuired (in MB)',
     Maintainer: `The person or entity who maintains this package`,
     'Changed-By': 'The person who created this build of the package',
     Homepage: 'Link to the homepage for the package',
@@ -43,23 +49,6 @@ const altDocsFields = {
   revision: 'version',
 }
 
-
-const Field = ({ field, fields, defaults, data, setData }) => (
-  <StringInput
-    key={field}
-    placeholder={defaults[field]}
-    label={field}
-    labelBL={fields[field]}
-    labelBR={<Docs field={field.toLowerCase()} />}
-    valid={() => true}
-    current={data[field]}
-    update={(val) => {
-      const newData = { ...data }
-      newData[field] = val
-      setData(newData)
-    }}
-  />
-)
 /**
  * The actual component, in case we want to extract it for re-use later
  */
@@ -99,7 +88,7 @@ const CreatePackage = () => {
     const getDefaults = async () => {
       let result
       try {
-        result = await api.getClientPackageDefaults('deb')
+        result = await api.getClientRepoPackageDefaults('deb')
       } catch (err) {
         if (err) console.log(err)
       }
@@ -117,7 +106,7 @@ const CreatePackage = () => {
    */
   const buildPackage = async () => {
     setLoadingStatus([true, 'Contacting Morio Management API'])
-    const [body, status] = await api.buildClientPackage('deb', data)
+    const [body, status] = await api.buildClientRepoPackage('deb', data)
     if (status === 201) {
       setResult(body)
       // Force update of the revision
@@ -209,7 +198,7 @@ const CreatePackage = () => {
     <>
       <h2 className="flex flex-row justify-between items-center">
         <span>
-          Create <b>.deb</b> Morio client package
+          Create a <b>.deb</b> repo setup package
         </span>
         <div className="flex flex-row gap-2">
           <button className="btn btn-primary btn-outline btn-sm" onClick={empty}>
@@ -225,20 +214,51 @@ const CreatePackage = () => {
       </h2>
       <Tabs tabs="basics,advanced">
         <Tab tabId="basics" key="basics">
-          {Object.keys(fields.basics).slice(0,1).map(field => (
-            <Field key={field} fields={fields.basics} {...{ defaults, data, setData, field }} />
-          ))}
-
           <div className="grid grid-cols-2 gap-4 items-end">
-            {Object.keys(fields.basics).slice(1,3).map(field => (
-              <Field key={field} fields={fields.basics} {...{ defaults, data, setData, field }} />
+            {Object.keys(fields.basics).map((field) => (
+              <StringInput
+                key={field}
+                placeholder={defaults[field]}
+                label={field}
+                labelBL={fields.basics[field]}
+                labelBR={<Docs field={field.toLowerCase()} />}
+                valid={() => true}
+                current={data[field]}
+                update={(val) => {
+                  const newData = { ...data }
+                  newData[field] = val
+                  setData(newData)
+                }}
+              />
             ))}
           </div>
         </Tab>
         <Tab tabId="advanced" key="advanced">
           <div className="grid grid-cols-2 gap-4 items-end">
-            {Object.keys(fields.advanced).map(field => (
-              <Field key={field} fields={fields.advanced} {...{ defaults, data, setData, field }} />
+            <StringInput
+              placeholder={defaults.Source}
+              label="Source Package Name"
+              labelBL="The name to use for the source package"
+              labelBR={<Docs field="source" />}
+              valid={() => true}
+              current={data.Source}
+              update={(Source) => setData({ ...data, Source })}
+            />
+            {Object.keys(fields.advanced).map((field) => (
+              <StringInput
+                key={field}
+                placeholder={defaults[field]}
+                label={field}
+                labelBL={fields.advanced[field]}
+                labelBR={<Docs field={field.toLowerCase()} />}
+                valid={() => true}
+                current={data[field]}
+                update={(val) => {
+                  const newData = { ...data }
+                  newData[field] = val
+                  setData(newData)
+                }}
+              />
             ))}
           </div>
           <TextInput
@@ -320,7 +340,7 @@ const CreatePackage = () => {
       </Tabs>
       <p className="text-center">
         <button onClick={buildPackage} className="btn btn-primary btn-lg">
-          Build .deb Package
+          Build Debian Repo Setup Package
         </button>
       </p>
     </>
@@ -343,7 +363,7 @@ export default DebPage
 
 export const getStaticProps = () => ({
   props: {
-    title: 'Morio Debian client builder',
-    page: ['tools', 'pkgs', 'deb'],
+    title: 'Debian repo setup builder',
+    page: ['actions', ['pkgs', 'Build Packages'], ['deb', 'Morio Debuan Repo Setup Builder']],
   },
 })
