+++ ui/pages/actions/pkgs/rpm.mjs
